Showcase Using Generic Collection Interfaces
Updated the SendEmail method of the vendor class so that its first parameter is ICollection<Vendor> instead of IList<Vendor>. This allows us to pass in any collection type the implements ICollection<T>.

Made several updates to the VendorRepository class to replace specific collection classes (in return types and parameters) to more generic interfaces (such as IEnumerable<T>). Also removed a couple of methods that became obsolete once the parameters and return types of certain methods were replaced with interfaces.